# spec/ais-api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Integrity Protocol — AIS API
  description: Read-side wire protocol for querying an agent's Agent Integrity Score (AIS), on-chain primitive addresses, compliance status, market state, and wallet balances. This is the versioned, externally-supported surface — /v1/* is additive-only for its lifetime; any field rename, type change, or removal requires a /v2/* prefix. See spec/ais-api/v1/README.md for the integration guide and spec/ais-api/CHANGELOG.md for the version history.
  license:
    name: ''
  version: 1.0.0
paths:
...
  /v1/agent/{id}/wallet:
    get:
      tags:
      - wallet
      operationId: get_wallet
      parameters:
      - name: id
        in: path
        description: Agent DID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: $ITK balance + open market positions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          description: Unknown DID
...
components:
  schemas:
...
    AllowanceDto:
      type: object
      required:
      - agent
      - limit
      - spent
      - status
      properties:
        agent:
          type: string
        limit:
          type: string
        spent:
          type: number
          format: double
        status:
          type: string
...
    TransactionDto:
      type: object
      required:
      - id
      - type
      - asset
      - amount
      - agent
      - status
      - time
      properties:
        agent:
          type: string
        amount:
          type: string
        asset:
          type: string
        id:
          type: string
        status:
          type: string
        time:
          type: string
        type:
          type: string
        usd:
          type:
          - string
          - 'null'
...
    WalletPositionDto:
      type: object
      required:
      - market_address
      - question
      - outcome_index
      - amount
      - market_resolved
      properties:
        amount:
          type: string
        market_address:
          type: string
        market_resolved:
          type: boolean
        outcome_index:
          type: integer
          format: int32
          minimum: 0
        question:
          type: string
        won:
          type:
          - boolean
          - 'null'
          description: '`Some(bool)` only once the market has resolved; `None` while still open.'
    WalletResponse:
      type: object
      required:
      - agent_id
      - sovereign_agent
      - itk_balance
      - open_positions
      properties:
        agent_id:
          type: string
        allowances:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/AllowanceDto'
        itk_balance:
          type: string
          description: Real `IntegrityToken.balanceOf(sovereignAgent)` read, decimal string.
        open_positions:
          type: array
          items:
            $ref: '#/components/schemas/WalletPositionDto'
          description: |-
            Unclaimed positions (amount > 0, `claimed == false`) across every market in the
            markets cache, cross-referenced via a real `getPosition` read per market. Bounded
            by the current market count — fine at this scale, would want indexing if the
            market count grows into the hundreds+.
        sovereign_agent:
          type: string
        transaction_history:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/TransactionDto'
          description: |-
            Transfer/stake/payout history requires indexing on-chain events (`Transfer`,
            `PositionEntered`, `PayoutClaimed`, ...), which this pass does not build. `null`,
            never a fabricated transaction list. See `docs/wiki/entities/integrity-oracle.md`.
...